add dish management

+++ sky-server/src/main/java/com/sky/service/impl/CategoryServiceImpl.java
@@ -2,12 +2,15 @@ package com.sky.service.impl;
 
 import com.github.pagehelper.Page;
 import com.github.pagehelper.PageHelper;
+import com.sky.constant.MessageConstant;
 import com.sky.constant.StatusConstant;
 import com.sky.context.BaseContext;
 import com.sky.dto.CategoryDTO;
 import com.sky.dto.CategoryPageQueryDTO;
 import com.sky.entity.Category;
+import com.sky.exception.DeletionNotAllowedException;
 import com.sky.mapper.CategoryMapper;
+import com.sky.mapper.DishMapper;
 import com.sky.result.PageResult;
 import com.sky.service.CategoryService;
 import lombok.extern.slf4j.Slf4j;
@@ -27,6 +30,8 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Autowired
     private CategoryMapper categoryMapper;
+    @Autowired
+    private DishMapper dishMapper;
 
     /**
      * 新增分类
@@ -70,8 +75,14 @@ public class CategoryServiceImpl implements CategoryService {
     @Override
     public void deleteById(Long id) {
         //判断当前分类是否关联了菜品
+        // 查菜品表此分类id下有几个条目
+        Integer dishCount = dishMapper.countByCategoryId(id);
+        // 如有，则该分给还有菜品与之关联，不能删，抛异常
+        if (dishCount > 0) {
+            throw new DeletionNotAllowedException(MessageConstant.CATEGORY_BE_RELATED_BY_DISH);
+        }
 
-        //若关联，抛异常
+        //TODO 判断当前分类是否关联了套餐
 
         //传参id删除该分类
         categoryMapper.delete(id);
